Add a test for session header encoding / decoding
